savings plan & non cumulative cost fix

+++ src/etl/shared.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 
 
 @dataclass
@@ -11,7 +12,7 @@ class Period:
 @dataclass
 class Quantity:
     unit: str
-    value: int
+    value: Decimal
 
 
 @dataclass
